validate destroy

+++ app/controllers/api_controller.rb
@@ -42,6 +42,10 @@ class ApiController < ApplicationController
   # Destroy Order product http://localhost:3000/api/destroy_order_product?order_product[order_id]=1&order_product[product_id]=1
   def destroy_order_product
     @order_product = OrderProduct.find_by order_product_params
+    unless @order_product
+      render json: { result: true, object: 'No existe' }
+      return;
+    end
 
     if @order_product.destroy
       render json: { result: true, object: @order_product }
